Some restructuring to address PR comments RE docstrings, threading, etc
