+++ dolosse/run_control/cli/README.md
@@ -1,6 +1,7 @@
 # Kafka interface to MIDAS (Python)
 
-A Python interface script for [MIDAS](https://midas.triumf.ca/MidasWiki/index.php/Main_Page) experiments that sends json format messages to a Kafka topic for control 
+A Python interface script for experiment run control that sends json format messages to a Kafka topic and reads out
+from a topic to display DAQ status information.
 
 
 ## Contents
@@ -14,7 +15,7 @@ Examples of the json formats for the message categories are included.
 
 
 ```bash
-python kafka_pyinq.py  # PyInquirer-based interactive CLI that accepts all inputs and shows DAQ status
+python kafka_run_control_interface.py --yaml-config [absolute path to yaml config]  # PyInquirer-based interactive CLI that accepts all inputs and shows DAQ status
 ```
 
 The interface has a manual refresh option for DAQ status information. 
@@ -22,7 +23,6 @@ The interface has a manual refresh option for DAQ status information.
 
 >The associated .yaml file must include configuration information for Kafka brokers
 
-**Note:** The script produces status messages to stderr , so pipe those to a log file
 ## Requirements
 
 Confluent kafka, PyInquirer, yaml
